# optimize NewFromString a bit