--- a/decimal_bench_test.go
+++ b/decimal_bench_test.go
@@ -1,6 +1,7 @@
 package decimal
 
 import (
+	"fmt"
 	"math"
 	"math/rand"
 	"sort"
@@ -181,5 +182,25 @@ func BenchmarkDecimal_IsInteger(b *testing.B) {
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
 		d.IsInteger()
+	}
+}
+
+func BenchmarkDecimal_NewFromString(b *testing.B) {
+	count := 72
+	prices := make([]string, 0, count)
+	for i := 1; i <= count; i++ {
+		prices = append(prices, fmt.Sprintf("%d.%d", i*100, i))
+	}
+
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		for _, p := range prices {
+			d, err := NewFromString(p)
+			if err != nil {
+				b.Log(d)
+				b.Error(err)
+			}
+		}
 	}
 }