Hi, while looking through the code, I noticed a string split being used and thought I'd make a pr for optimizing the parsing of a string a bit.

original:
```
goos: darwin
goarch: amd64
pkg: github.com/shopspring/decimal
BenchmarkDecimal_NewFromString_large_number
BenchmarkDecimal_NewFromString_large_number-12    	   23740	     50176 ns/op	   14400 B/op	     432 allocs/op
PASS

goos: darwin
goarch: amd64
pkg: github.com/shopspring/decimal
BenchmarkDecimal_NewFromString
BenchmarkDecimal_NewFromString-12    	   48504	     24745 ns/op	    8064 B/op	     360 allocs/op
PASS
```

new numbers 

for large numbers:
```
goos: darwin
goarch: amd64
pkg: github.com/shopspring/decimal
BenchmarkDecimal_NewFromString_large_number
BenchmarkDecimal_NewFromString_large_number-12    	   27516	     43336 ns/op	   12096 B/op	     360 allocs/op
PASS
```

optimization for smaller numbers:
```
goos: darwin
goarch: amd64
pkg: github.com/shopspring/decimal
BenchmarkDecimal_NewFromString
BenchmarkDecimal_NewFromString-12    	  102325	     11520 ns/op	    3456 B/op	     216 allocs/op
PASS
```