@mwoss 2021-01-10
LGTM too, thanks @nknordeen . 
There is a space for simplification of branching in this init method, but we can work on this later. :)